feat: post thread view modal integrated with send modal states
